Semesterly: Refactor Semesterly component
