+++ static/js/redux/ui/alert_box.jsx
@@ -137,7 +137,7 @@ class AlertBox extends React.Component {
 
   /**
    * Show the alert in the page
-   * @param  {string} message
+   * @param  message
    * @param  {Object} options
    * @return {void}
    */
