+++ static/js/redux/ui/containers/semesterly_container.jsx
@@ -14,8 +14,6 @@ GNU General Public License for more details.
 
 import { connect } from "react-redux";
 import Semesterly from "../semesterly";
-import { saveTimetable } from "../../actions/user_actions";
-import { setActiveTimetable } from "../../actions/timetable_actions";
 import { getTimetables, getActiveTimetableCourses } from "../../state";
 
 const mapStateToProps = (state) => {
@@ -41,13 +39,9 @@ const mapStateToProps = (state) => {
     explorationModalIsVisible: state.explorationModal.isVisible,
     dataLastUpdated: state.school.dataLastUpdated,
     PgCount: timetables.length,
-    PgActive: active,
   };
 };
 
-const SemesterlyContainer = connect(mapStateToProps, {
-  saveTimetable,
-  setPgActive: setActiveTimetable,
-})(Semesterly);
+const SemesterlyContainer = connect(mapStateToProps, {})(Semesterly);
 
 export default SemesterlyContainer;
